updated SmaAnalyseWorker

+++ ExchangeAnalyticsService/ExchCommonLib/Classes/Candle.cs
@@ -9,10 +9,10 @@ namespace ExchCommonLib.Classes
     public class Candle
     {
         public DateTime Date { get; set; }
-        public float Open { get; set; }
-        public float High { get; set; }
-        public float Low { get; set; }
-        public float Close { get; set; }
+        public double Open { get; set; }
+        public double High { get; set; }
+        public double Low { get; set; }
+        public double Close { get; set; }
         public ulong Volume { get; set; }
 
         public CandlesInterval Interval { get; set; }
